Project/IT/backend/cpms/requirements/base.txt
Django==4.1.5
djangorestframework==3.14.0
django-filter==22.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework-simplejwt==5.2.2
requests==2.28.2